### 2022-09-28 4e87a1a
[SKIP] version bump 0.1.10 -> 0.1.11
Merge pull request #28 from menoliu/fix-smfret

Fix `smfret` output formatting
