### 2022-09-29 e3a5538
[SKIP] version bump 0.1.11 -> 0.1.12
Merge pull request #29 from joaomcteixeira/paper_minor

minor corrections to the paper
