+++ CHANGELOG.rst
@@ -2,6 +2,9 @@
 Changelog
 =========
 
+v0.1.12 (2022-09-29)
+------------------------------------------------------------
+
 * Minor edits for the paper
 
 v0.1.11 (2022-09-28)
