### 2022-09-30 ee0314e
[SKIP] version bump 0.1.12 -> 0.1.13
Merge pull request #30 from menoliu/mod-saxs

Modify `saxs` module output to be uniform
