+++ CHANGELOG.rst
@@ -2,6 +2,9 @@
 Changelog
 =========
 
+v0.1.13 (2022-09-30)
+------------------------------------------------------------
+
 * Modify SAXS module to follow "format" in output
 
 v0.1.12 (2022-09-29)
