### 2022-10-17 #4861 @editorialbot
[REVIEW]: SPyCi-PDB: A modular command-line interface for back-calculating experimental datatypes of protein structures
**Submitting author:** @menoliu (<a href="http://orcid.org/0000-0002-8357-8507">Zi Hao Liu</a>)
**Repository:** https://github.com/julie-forman-kay-lab/SPyCi-PDB
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.5
