### 2022-11-28 julie-forman-kay-lab/SPyCi-PDB#31 @lohedges
[JOSS Review] Unable to run spycipdb without "optional" dependency CSpred
[Re: JOSS review](https://github.com/openjournals/joss-reviews/issues/4861)

From the [documentation](https://spyci-pdb.readthedocs.io/en/stable/installation.html#installing-third-party-software) it is implied that `UCBShift` is an optional dependency, i.e. it is not mandatory to run the `spycipdb` command-line tool. However, having installed followed the instructions to create a environment with `IDPConformerGenerator` and `SPyCi-PDB` installed, I get the following error:

```sh
$ spycipdb
Traceback (most recent call last):
  File "/home/lester/.conda/envs/spycipdb/bin/spycipdb", line 33, in <module>
    sys.exit(load_entry_point('spycipdb', 'console_scripts', 'spycipdb')())
  File "/home/lester/.conda/envs/spycipdb/bin/spycipdb", line 25, in importlib_load_entry_point
    return next(matches).load()
  File "/home/lester/.conda/envs/spycipdb/lib/python3.10/importlib/metadata/__init__.py", line 171, in load
    module = import_module(match.group('module'))
  File "/home/lester/.conda/envs/spycipdb/lib/python3.10/importlib/__init__.py", line 126, in import_module
  File "<frozen importlib._bootstrap>", line 1050, in _gcd_import
  File "<frozen importlib._bootstrap>", line 1027, in _find_and_load
  File "<frozen importlib._bootstrap>", line 1006, in _find_and_load_unlocked
  File "<frozen importlib._bootstrap>", line 688, in _load_unlocked
  File "<frozen importlib._bootstrap_external>", line 883, in exec_module
  File "<frozen importlib._bootstrap>", line 241, in _call_with_frames_removed
  File "/home/lester/Downloads/JOSS/SPyCi-PDB/src/spycipdb/clis/cli.py", line 18, in <module>
    from spycipdb.clis import (
  File "/home/lester/Downloads/JOSS/SPyCi-PDB/src/spycipdb/clis/cli_cs.py", line 55, in <module>
    from spycipdb.components.helpers import calc_sing_pdb
  File "/home/lester/Downloads/JOSS/SPyCi-PDB/src/spycipdb/components/helpers.py", line 24, in <module>
    from CSpred import calc_sing_pdb  # noqa: E402, F401
ModuleNotFoundError: No module named 'CSpred'
```

If I follow the [installation instructions](https://spyci-pdb.readthedocs.io/en/stable/installation.html#ucbshift) for `UCBShift`, then reinstall `spycipdb`, then it works as expected.

in [src/components/helpers.py](https://github.com/julie-forman-kay-lab/SPyCi-PDB/blob/ee0314e982c6fcff8443aa3a50a678f8ae3e66dc/src/spycipdb/components/helpers.py#L24) you try to directly import from `CSpred`, with no checks to see whether it is installed. Perhaps this should be wrapped in a `try/except` block, so that functionality can be _optionally_ exposed if it is present.
