### 2022-11-28 73c026b
catch missing UCBShift imports
