### 2022-11-28 julie-forman-kay-lab/SPyCi-PDB#31 @menoliu
Thanks Lester, I have addressed this and updated the documentation in #32
