### 2022-11-29 julie-forman-kay-lab/SPyCi-PDB#33 @lohedges
[JOSS Review] Installation and packaging
[Re: JOSS Review](https://github.com/openjournals/joss-reviews/issues/4861)

The installation process is a little confusing and somewhat non-standard. (It seems to be primarily targeted towards developers.) Although you state that `idpconfgen` is required as a dependency at the outset, it is not installed alongside `spycipdb`. The installation instructions start by geting you to set up an environment for `spycipdb`, e.g.:

```sh
conda env create -f requirements.yml
conda activate spycipdb
```

You then install `spycipdb` into this environment in development mode:

```sh
python setup.py develop --no-deps
```

At this point you have an environment with a _non-working_ `spycipdb` install, i.e:

```sh
spycipdb
Traceback (most recent call last):
  File "/home/lester/.conda/envs/spycipdb/bin/spycipdb", line 33, in <module>
    sys.exit(load_entry_point('spycipdb', 'console_scripts', 'spycipdb')())
  File "/home/lester/.conda/envs/spycipdb/bin/spycipdb", line 25, in importlib_load_entry_point
    return next(matches).load()
  File "/home/lester/.conda/envs/spycipdb/lib/python3.10/importlib/metadata/__init__.py", line 171, in load
    module = import_module(match.group('module'))
  File "/home/lester/.conda/envs/spycipdb/lib/python3.10/importlib/__init__.py", line 126, in import_module
  File "<frozen importlib._bootstrap>", line 1050, in _gcd_import
  File "<frozen importlib._bootstrap>", line 1027, in _find_and_load
  File "<frozen importlib._bootstrap>", line 1006, in _find_and_load_unlocked
  File "<frozen importlib._bootstrap>", line 688, in _load_unlocked
  File "<frozen importlib._bootstrap_external>", line 883, in exec_module
  File "<frozen importlib._bootstrap>", line 241, in _call_with_frames_removed
  File "/home/lester/Downloads/JOSS/SPyCi-PDB/src/spycipdb/clis/cli.py", line 18, in <module>
    from spycipdb.clis import (
  File "/home/lester/Downloads/JOSS/SPyCi-PDB/src/spycipdb/clis/cli_cs.py", line 52, in <module>
    from idpconfgen.libs.libmulticore import pool_function
ModuleNotFoundError: No module named 'idpconfgen'
```

It is only _after_ this that you have instructions for installing the required `idpconfgen`, after which you need to reinstall `spycipdb` anyway.

I understand that you may wish to separate the two packages to allow them to be developed and updated independently, but if `idpconfgen` is strictly required by `spycipdb`, then perhaps it would be good to install it at the time that the `spycipdb` environment is resolved. This could be done as follows, which would still result in a development install of `idpconfgen`:

```python
diff --git a/requirements.yml b/requirements.yml
index f7c96bb..bcb00a0 100644
--- a/requirements.yml
+++ b/requirements.yml
@@ -4,6 +4,7 @@ channels:
     - defaults
 dependencies:
     - python>=3.8,<4
+    - git
     - numpy>=1.22.4,<2
     - pandas>=1.2.4,<2
     - numba>=0.53.0,<1
@@ -11,4 +12,6 @@ dependencies:
     - scipy>=1.8.1,<2
     - matplotlib>=3.5.2,<4
     - pip
+    - pip:
+      - "--editable=git+https://github.com/julie-forman-kay-lab/IDPConformerGenerator.git#egg=idpconfgen"
     - tox>=3.25.0,<4
```

Another issue with your installation method is that you are re-solving the environment when adding additional (optional) dependencies, so the Python version for the resulting environment is ultimately determined by the last thing that is installed. For example, after installing only `spycipdb` and `idpconfgen` I end up with a Python 3.10 environment:

```sh
python --version
Python 3.10.8
```

However, after following the instructions to install `UCBShift`, this is downgraded to 3.8.15:

```
python --version
Python 3.8.15
```

This is why you then need to reinstall both `idpconfgen` and `spycipdb`, since they reside in the `site-packages` directory for the _original_ Python variant, i.e.:

```sh
ls -l ~/.conda/envs/spycipdb/lib/python3.10/site-packages
total 16K
drwxr-sr-x 2 lester lester 4.0K Nov 29 12:39 idpconfgen-0.6.16.dist-info
-rw-r--r-- 1 lester lester   42 Nov 29 12:39 easy-install.pth
-rw-r--r-- 1 lester lester   57 Nov 29 12:39 __editable__.idpconfgen-0.6.16.pth
-rw-r--r-- 1 lester lester   45 Nov 29 12:39 spycipdb.egg-link
```

After re-installing, they are now in `/.conda/envs/spycipdb/lib/python3.8/site-packages`, as required.

Although your paper specifies support for Python >= 3.8, < 4.0, the user doesn't have control over the variant that they end up with, i.e. it would be dependent on the optional packages that they install (and potentially the OS too). This would make it hard for a user to install your package alongside other tools within a larger environment, where they might require a specific Python variant for other things (or their own scripts) to work. This might not be an issue if you only intend `spycipdb` to be used in a stand-alone fashion, but would make it harder to be used within broader ecosystem.

Since your packages have minimal and standard dependencies I would advise creating packages for them. These could easily be built during your CI and directly uploaded to appropriate package repositories through which they could be distributed, e.g.  `PyPi` and the Anaconda Cloud (via `conda-forge`). This would make `spycipdb` and `idpconfgen` accessible to a much wider range of users, greatly simplify the installation process, and make it easier to use within a larger ecosystem of packages.
