### 2022-11-30 julie-forman-kay-lab/SPyCi-PDB#34 @lohedges
[JOSS Review] Misleading CSpred error message
[Re: JOSS Review](https://github.com/openjournals/joss-reviews/issues/4861)

Not directly related to `spycipdb`, but if `UCBShift` is installed, using your [fork](https://github.com/menoliu/CSpred) as described in the instructions, then a misleading error message appears when running the `spycipdb` command, i.e.:

```
spycipdb

which: no reduce in ...

usage:
    $ spycipdb pre <PDB-FILES> [--exp-file]
    $ spycipdb pre <PDB-FILES> [--exp-file] [--output] [--ncores]
    ...
```

I think this is because of [this](https://github.com/menoliu/CSpred/blob/fce6d51fa57d136b1c290bbb6cb9f45ce845a39f/toolbox.py#L23-L27) try/except block, which is run as a subprocess but doesn't capture standard output/error.

I'm happy to report this as an issue on the main `CSpred` repository, but it might be an easy fix in your fork that would avoid confusing your users. In my case, before I figured out where the error message was coming from (by grepping the parent directory of my clones for `reduce`), I thought that something was wrong with my `spycipdb` installation.
