### 2022-12-01 julie-forman-kay-lab/SPyCi-PDB#35 @lohedges
[JOSS Review] Examples and input data validation
[Re: JOSS Review](https://github.com/openjournals/joss-reviews/issues/4861)

It would be nice if the [example usage](https://spyci-pdb.readthedocs.io/en/stable/usage.html#basic-usage-examples) section could be expanded to provide a little more context and to walk a new user through exactly what is going on. Perhaps explaining what the different back calculators can calculate, e.g. for those not familiar with all of them. It would also be nice if there were some examples of what could be determined from the output, like the plot used in your paper.

It also wasn't clear exactly what would be allowed for certain input, e.g. the `pdb_files` positional argument:

```
positional arguments:
  pdb_files             Paths to PDB files in the disk. Accepts a TAR file.
```

This implies that it requires a `TAR` file, i.e. a `.tar` extension. However, the repository and example use a `.tar.xz` file. Running the example as is:

```
spycipdb pre ./drksh3_csss_100.tar.xz -e ./drksh3_exp_data/drksh3_PRE.txt -n
```

gives:

```
which: no reduce in (/home/lester/.conda/envs/spycipdb/bin:/opt/miniconda3/condabin:/usr/local/gromacs/bin:/usr/local/bin:/usr/local/sbin:/usr/bin:/opt/cuda/bin:/opt/cuda/nsight_compute:/opt/cuda/nsight_systems/bin:/usr/lib/jvm/default/bin:/usr/bin/site_perl:/usr/bin/vendor_perl:/usr/bin/core_perl)
[2022-12-01 12:50:18,333]Reading Input Paths:
[2022-12-01 12:50:18,334]Logging Non-Existent Files:
[2022-12-01 12:50:18,334]
[2022-12-01 12:50:18,334]    done
[2022-12-01 12:50:18,334]    done
[2022-12-01 12:50:18,334]Back Calculating Using 19 Workers:
   0.0% 0/0 on 19 cpus |-----------------------------------------------------------------------|
[2022-12-01 12:50:18,359]    done
[2022-12-01 12:50:18,359]Writing Output Onto Disk:
[2022-12-01 12:50:18,360]    done
[2022-12-01 12:50:18,360]    finished properly
```

Looking at the output I only see a `format` entry in the JSON, e.g.:

```
{
    "format": {
        "res1": [
            2,
            2,
            2,
            2,
            2,
            2,
            2,
            2,
            ...
}
```

Presumably, the `0.0% 0/0` in the output means that it hasn't actually done anything! Indeed, if I use `unxz` on the archive, it then works:

```
unxz drksh3_csss_100.tar.xz

spycipdb pre ./drksh3_csss_100.tar -e ./drksh3_exp_data/drksh3_PRE.txt -n
which: no reduce in (/home/lester/.conda/envs/spycipdb/bin:/opt/miniconda3/condabin:/usr/local/gromacs/bin:/usr/local/bin:/usr/local/sbin:/usr/bin:/opt/cuda/bin:/opt/cuda/nsight_compute:/opt/cuda/nsight_systems/bin:/usr/lib/jvm/default/bin:/usr/bin/site_perl:/usr/bin/vendor_perl:/usr/bin/core_perl)
[2022-12-01 13:04:38,390]Reading Input Paths:
[2022-12-01 13:04:38,396]    done
[2022-12-01 13:04:38,396]Back Calculating Using 19 Workers:
 100.0% 100/100 on 19 cpus |███████████████████████████████████████████████████████████████████████|
[2022-12-01 13:04:39,591]    done
[2022-12-01 13:04:39,591]Writing Output Onto Disk:
[2022-12-01 13:04:39,596]    done
[2022-12-01 13:04:39,597]    finished properly
```

Given that both runs state that they _finished properly_ it is rather misleading. It should be simple to validate that some conformers have been loaded and give an appropriate message if not. I would also imagine that it should be possible to support a range of archive formats, e.g. `.tar`, `.tar.gz`, `.tar.xz`, `.zip`, so that the user doesn't need to convert to the appropriate format. Unarchiving does work, i.e. passing the path to a directory including the PDB files, but _only_ if the file extensions are `.pdb`. Presumably this will generally be the case, but it might be nice to document this. (This would also be caught by making sure that conformers have been loaded.)

I also played around with the different back calculators using the example input templates in `example/drksh3_exp_data`. While you describe the required input formatting for each template, there doesn't seem to be any validation of the input. For example, I accidentally passed the `PRE` template when running the `NOE` back calculator and got:

```
spycipdb noe ./drksh3_csss_100.tar -e ./drksh3_exp_data/drksh3_PRE.txt -n

Traceback (most recent call last):
  File "/home/lester/.conda/envs/spycipdb/bin/spycipdb", line 33, in <module>
    sys.exit(load_entry_point('spycipdb', 'console_scripts', 'spycipdb')())
  File "/home/lester/Downloads/JOSS/SPyCi-PDB/src/spycipdb/clis/cli.py", line 95, in maincli
    cmd.func(**vars(cmd))
  File "/home/lester/Downloads/JOSS/SPyCi-PDB/src/spycipdb/clis/cli_noe.py", line 134, in main
    _output['format'] = get_exp_format_noe(exp_file)
  File "/home/lester/Downloads/JOSS/SPyCi-PDB/src/spycipdb/core/parsers.py", line 12, in get_exp_format_noe
    format['atom1_multiple_assignments'] = exp.atom1_multiple_assignments.values.tolist()  # noqa: E501
  File "/home/lester/.conda/envs/spycipdb/lib/python3.8/site-packages/pandas/core/generic.py", line 5902, in __getattr__
    return object.__getattribute__(self, name)
AttributeError: 'DataFrame' object has no attribute 'atom1_multiple_assignments'
```

If possible, it would be nice to do some simple validation so that the user can quickly understand why their input is wrong, e.g. wrong number of records per line, wrong type for a given record, etc. Perhaps there is sufficient overlap between the input formats that this is difficult, but it might be possible to make them unique, e.g. by re-ordering the records, or adding a header.

As a test I also modified the input files (sample templates) so that they had entries corresponding to atoms/residues that were not in the PDB data set, e.g. wrong names, residue/atom numbers outside of the PDB range, or negative. In this case things crashed spectacularly:

```
spycipdb pre ./drksh3_csss_100.tar -e ./drksh3_exp_data/drksh3_PRE.txt -n                         ✘ 1 main ✱ ◼
which: no reduce in (/home/lester/.conda/envs/spycipdb/bin:/opt/miniconda3/condabin:/usr/local/gromacs/bin:/usr/local/bin:/usr/local/sbin:/usr/bin:/opt/cuda/bin:/opt/cuda/nsight_compute:/opt/cuda/nsight_systems/bin:/usr/lib/jvm/default/bin:/usr/bin/site_perl:/usr/bin/vendor_perl:/usr/bin/core_perl)
[2022-12-01 13:14:53,686]Reading Input Paths:
[2022-12-01 13:14:53,694]    done
[2022-12-01 13:14:53,695]Back Calculating Using 19 Workers:
   0.0%   0/100 on 19 cpus |-----------------------------------------------------------------------|[2022-12-01 13:14:53,727]    saved ERROR REPORT: /home/lester/Downloads/JOSS/SPyCi-PDB/example/ROC_-3824236825196499224_1669900493727114433.rpr_on_crash
[2022-12-01 13:14:53,730]    saved ERROR REPORT: /home/lester/Downloads/JOSS/SPyCi-PDB/example/ROC_-8503020282077084117_1669900493730592498.rpr_on_crash
[2022-12-01 13:14:53,731]    saved ERROR REPORT: /home/lester/Downloads/JOSS/SPyCi-PDB/example/ROC_3085658112409125258_1669900493730844682.rpr_on_crash
[2022-12-01 13:14:53,731]    saved ERROR REPORT: /home/lester/Downloads/JOSS/SPyCi-PDB/example/ROC_8087155530390538253_1669900493730845716.rpr_on_crash
[2022-12-01 13:14:53,731]    saved ERROR REPORT: /home/lester/Downloads/JOSS/SPyCi-PDB/example/ROC_-6059140792673140665_1669900493730858581.rpr_on_crash
[2022-12-01 13:14:53,731]    saved ERROR REPORT: /home/lester/Downloads/JOSS/SPyCi-PDB/example/ROC_-1671896391808535828_1669900493730884456.rpr_on_crash
[2022-12-01 13:14:53,731]    saved ERROR REPORT: /home/lester/Downloads/JOSS/SPyCi-PDB/example/ROC_-7084764236959228053_1669900493731031247.rpr_on_crash
[2022-12-01 13:14:53,731]    saved ERROR REPORT: /home/lester/Downloads/JOSS/SPyCi-PDB/example/ROC_2052837033974899781_1669900493731032292.rpr_on_crash
[2022-12-01 13:14:53,731]    saved ERROR REPORT: /home/lester/Downloads/JOSS/SPyCi-PDB/example/ROC_7265407114291332825_1669900493731150448.rpr_on_crash
[2022-12-01 13:14:53,731]    saved ERROR REPORT: /home/lester/Downloads/JOSS/SPyCi-PDB/example/ROC_-946253142379564336_1669900493731150538.rpr_on_crash
[2022-12-01 13:14:53,731]    saved ERROR REPORT: /home/lester/Downloads/JOSS/SPyCi-PDB/example/ROC_-11073480369654686_1669900493731330557.rpr_on_crash

[2022-12-01 13:14:53,732]    saved ERROR REPORT: /home/lester/Downloads/JOSS/SPyCi-PDB/example/ROC_873434531058914103_1669900493731974534.rpr_on_crash
[2022-12-01 13:14:53,732]    saved ERROR REPORT: /home/lester/Downloads/JOSS/SPyCi-PDB/example/ROC_8738648456016735003_1669900493732087885.rpr_on_crash
[2022-12-01 13:14:53,732]    saved ERROR REPORT: /home/lester/Downloads/JOSS/SPyCi-PDB/example/ROC_4064632004247594731_1669900493732390903.rpr_on_crash
[2022-12-01 13:14:53,732]    saved ERROR REPORT: /home/lester/Downloads/JOSS/SPyCi-PDB/example/ROC_2178549454091240155_1669900493732476356.rpr_on_crash
[2022-12-01 13:14:53,732]    saved ERROR REPORT: /home/lester/Downloads/JOSS/SPyCi-PDB/example/ROC_2448285797439649646_1669900493732601141.rpr_on_crash
multiprocessing.pool.RemoteTraceback:
"""
Traceback (most recent call last):
  File "/home/lester/Downloads/JOSS/SPyCi-PDB/src/spycipdb/logger.py", line 117, in report_on_crash
    return func(*args, **kwargs)
  File "/home/lester/Downloads/JOSS/SPyCi-PDB/src/spycipdb/core/calculators.py", line 135, in calc_pre
    dv = atom1 - atom2
UnboundLocalError: local variable 'atom1' referenced before assignment

Traceback (most recent call last):
  File "/home/lester/.conda/envs/spycipdb/lib/python3.8/multiprocessing/pool.py", line 125, in worker
    result = (True, func(*args, **kwds))
  File "/home/lester/Downloads/JOSS/SPyCi-PDB/src/spycipdb/logger.py", line 146, in report_on_crash
    raise roc_error from err
spycipdb.core.exceptions.ReportOnCrashError: Crash reported to /home/lester/Downloads/JOSS/SPyCi-PDB/example/ROC_-3824236825196499224_1669900493727114433.rpr_on_crash.
"""

Traceback (most recent call last):
  File "/home/lester/.conda/envs/spycipdb/bin/spycipdb", line 33, in <module>
    sys.exit(load_entry_point('spycipdb', 'console_scripts', 'spycipdb')())
  File "/home/lester/Downloads/JOSS/SPyCi-PDB/src/spycipdb/clis/cli.py", line 95, in maincli
    cmd.func(**vars(cmd))
  File "/home/lester/Downloads/JOSS/SPyCi-PDB/src/spycipdb/clis/cli_pre.py", line 129, in main
    for result in execute_pool:
  File "/home/lester/Downloads/JOSS/IDPConformerGenerator/src/idpconfgen/libs/libmulticore.py", line 83, in pool_function
    yield next(imap)
  File "/home/lester/.conda/envs/spycipdb/lib/python3.8/multiprocessing/pool.py", line 868, in next
    raise value
spycipdb.core.exceptions.ReportOnCrashError: Crash reported to /home/lester/Downloads/JOSS/SPyCi-PDB/example/ROC_-3824236825196499224_1669900493727114433.rpr_on_crash.
```

The above would look a bit intimidating to a new user. It would be good if the template could be validated before proceeding, i.e. checking that it is self-consistent with the PDB files. You would then be able to give a meaningful error message such as "residue number X not found...".
