### 2023-01-16 julie-forman-kay-lab/SPyCi-PDB#33 @menoliu
Thanks for finding this bug Lester! It is also a good idea to see what I can do about package handling with `PyPi` but I will leave that improvement for a later date. I have solved this issue in #36
