### 2023-01-16 c632ffd
[SKIP] version bump 0.1.14 -> 0.1.15
Merge pull request #36 from menoliu/fix_installation

Fix installation
