+++ src/spycipdb/__init__.py
@@ -71,4 +71,4 @@ def count_string_formatters(s):
 
 source_folder = Path(__file__).absparent
 
-__version__ = '0.1.14'
+__version__ = '0.1.15'
