+++ src/spycipdb/clis/cli_pre.py
@@ -114,6 +114,12 @@ def main(
     
     log.info(T('reading input paths'))
     pdbs2operate, _istarfile = get_pdb_paths(pdb_files, tmpdir)
+    if len(pdbs2operate) == 0:
+        log.info(
+            'No .pdb files were found based on the input. Make sure the '
+            'folder/tarball contains .pdb files. Only .tar, .tar.xz, .tar.gz '
+            'tarballs are accepted.'
+            )
     log.info(S('done'))
     
     log.info(T(f'back calculating using {ncores} workers'))
