### 2023-01-16 f6ab8bd
catches incorrect formats for experimental files
