+++ src/spycipdb/core/parsers.py
@@ -7,12 +7,15 @@ def get_exp_format_noe(fexp):
     format = {}
     exp = pd.read_csv(fexp)
     
-    format['res1'] = exp.res1.values.astype(int).tolist()
-    format['atom1'] = exp.atom1.values.tolist()
-    format['atom1_multiple_assignments'] = exp.atom1_multiple_assignments.values.tolist()  # noqa: E501
-    format['res2'] = exp.res2.values.astype(int).tolist()
-    format['atom2'] = exp.atom2.values.tolist()
-    format['atom2_multiple_assignments'] = exp.atom2_multiple_assignments.values.tolist()  # noqa: E501
+    try:
+        format['res1'] = exp.res1.values.astype(int).tolist()
+        format['atom1'] = exp.atom1.values.tolist()
+        format['atom1_multiple_assignments'] = exp.atom1_multiple_assignments.values.tolist()  # noqa: E501
+        format['res2'] = exp.res2.values.astype(int).tolist()
+        format['atom2'] = exp.atom2.values.tolist()
+        format['atom2_multiple_assignments'] = exp.atom2_multiple_assignments.values.tolist()  # noqa: E501
+    except AttributeError:
+        return False
     
     return format
 
@@ -22,10 +25,13 @@ def get_exp_format_pre(fexp):
     format = {}
     exp = pd.read_csv(fexp)
     
-    format['res1'] = exp.res1.values.astype(int).tolist()
-    format['atom1'] = exp.atom1.values.tolist()
-    format['res2'] = exp.res2.values.astype(int).tolist()
-    format['atom2'] = exp.atom2.values.tolist()
+    try:
+        format['res1'] = exp.res1.values.astype(int).tolist()
+        format['atom1'] = exp.atom1.values.tolist()
+        format['res2'] = exp.res2.values.astype(int).tolist()
+        format['atom2'] = exp.atom2.values.tolist()
+    except AttributeError:
+        return False
     
     return format
 
@@ -35,8 +41,11 @@ def get_exp_format_smfret(fexp):
     format = {}
     exp = pd.read_csv(fexp)
     
-    format['res1'] = exp.res1.values.astype(int).tolist()
-    format['res2'] = exp.res2.values.astype(int).tolist()
-    format['scale'] = exp.scale.values.tolist()
+    try:
+        format['res1'] = exp.res1.values.astype(int).tolist()
+        format['res2'] = exp.res2.values.astype(int).tolist()
+        format['scale'] = exp.scale.values.tolist()
+    except AttributeError:
+        return False
     
     return format
