+++ src/spycipdb/clis/cli_pre.py
@@ -132,7 +132,7 @@ def main(
     
     _output = {}
     _output['format'] = get_exp_format_pre(exp_file)
-    if get_exp_format_pre(exp_file) == False:
+    if get_exp_format_pre(exp_file) is False:
         log.info(
             'Incorrect experimental file format for PRE subclient. '
             'Text file must have the following columns: '
