### 2023-01-17 4886a79
better way to communicate error message
