+++ src/spycipdb/clis/cli_saxs.py
@@ -136,7 +136,7 @@ def main(
     log.info(T('reading input paths'))
     pdbs2operate, _istarfile = get_pdb_paths(pdb_files, tmpdir)
     str_pdbpaths = [str(path) for path in pdbs2operate]
-    if len(pdbs2operate) == 0:
+    if len(pdbs2operate) == 0 or pdbs2operate == None:
         log.info(
             'No .pdb files were found based on the input. Make sure the '
             'folder/tarball contains .pdb files. Only .tar, .tar.xz, .tar.gz '
