+++ src/spycipdb/clis/cli_pre.py
@@ -114,7 +114,7 @@ def main(
     
     log.info(T('reading input paths'))
     pdbs2operate, _istarfile = get_pdb_paths(pdb_files, tmpdir)
-    if len(pdbs2operate) == 0:
+    if len(pdbs2operate) == 0 or pdbs2operate == None:
         log.info(
             'No .pdb files were found based on the input. Make sure the '
             'folder/tarball contains .pdb files. Only .tar, .tar.xz, .tar.gz '
@@ -132,13 +132,6 @@ def main(
     
     _output = {}
     _output['format'] = get_exp_format_pre(exp_file)
-    if get_exp_format_pre(exp_file) is False:
-        log.info(
-            'Incorrect experimental file format for PRE subclient. '
-            'Text file must have the following columns: '
-            )
-        log.info('res1,atom1,res2,atom2')
-        return
     for result in execute_pool:
         _output[result[0].stem] = result[1]
     log.info(S('done'))
