### 2023-01-17 440fae7
validate for residue numbers and atom names
