+++ src/spycipdb/clis/cli_jc.py
@@ -109,7 +109,7 @@ def main(
     
     log.info(T('reading input paths'))
     pdbs2operate, _istarfile = get_pdb_paths(pdb_files, tmpdir)
-    if len(pdbs2operate) == 0 or pdbs2operate == None:
+    if len(pdbs2operate) == 0 or pdbs2operate is None:
         log.info(
             'No .pdb files were found based on the input. Make sure the '
             'folder/tarball contains .pdb files. Only .tar, .tar.xz, .tar.gz '
