+++ src/spycipdb/clis/cli_smfret.py
@@ -123,7 +123,7 @@ def main(
     
     log.info(T('reading input paths'))
     pdbs2operate, _istarfile = get_pdb_paths(pdb_files, tmpdir)
-    if len(pdbs2operate) == 0 or pdbs2operate == None:
+    if len(pdbs2operate) == 0 or pdbs2operate is None:
         log.info(
             'No .pdb files were found based on the input. Make sure the '
             'folder/tarball contains .pdb files. Only .tar, .tar.xz, .tar.gz '
@@ -141,7 +141,7 @@ def main(
     execute_pool = pool_function(execute, pdbs2operate, ncores=ncores)
     
     _output = {}
-    _output['format'] = get_exp_format_smfret(exp_file)
+    _output['format'], _ = get_exp_format_smfret(exp_file, pdbs2operate[0])
     for results in execute_pool:
         _output[results[0].stem] = results[1]
     log.info(S('done'))
