+++ src/spycipdb/core/parsers.py
@@ -7,16 +7,25 @@ sample PDB input and the data in question.
 I.e. Sees if the experimental file has any negative values for residue
 numbers and if there are inconsistencies between atom names.
 """
+import numpy as np
 import pandas as pd
+from idpconfgen.libs.libstructure import Structure, col_name, col_resSeq
 
 from spycipdb.core.exceptions import SPyCiPDBException
 
 
-def get_exp_format_noe(fexp):
+def get_exp_format_noe(fexp, fpdb):
     """Get format from experimental template."""
+    errmsgs = []
     format = {}
     exp = pd.read_csv(fexp)
     
+    struc = Structure(fpdb)
+    struc.build()
+    struc_residues = struc.data_array[:, col_resSeq]
+    struc_atomnames = struc.data_array[:, col_name]
+    last_residue = int(struc.data_array[:, col_resSeq][-1])
+
     try:
         format['res1'] = exp.res1.values.astype(int).tolist()
         format['atom1'] = exp.atom1.values.tolist()
@@ -32,13 +41,53 @@ def get_exp_format_noe(fexp):
             )
         raise SPyCiPDBException(errmsg) from err
     
-    return format
+    for i, res1 in enumerate(format['res1']):
+        errmsg = False
+        atom1 = format['atom1'][i]
+        res2 = format['res2'][i]
+        atom2 = format['atom2'][i]
+        
+        if res1 <= 0 or res2 <= 0:
+            errmsg = 'res1/res2 cannot contain 0 or negative values'
+        elif res1 > last_residue or res2 > last_residue:
+            errmsg = 'res1/res2 cannot be greater than the maximum number of residues in your PDB structure.'  # noqa: E501
+        
+        try:
+            res1_idxs = np.where(struc_residues == str(res1))
+            res2_idxs = np.where(struc_residues == str(res2))
+            atom1_idxs = np.where(struc_atomnames == atom1)
+            atom2_idxs = np.where(struc_atomnames == atom2)
+            
+            one = np.isin(res1_idxs, atom1_idxs)
+            two = np.isin(res2_idxs, atom2_idxs)
+            
+            if True not in one:
+                errmsgs.append(f'{atom1} in PDB does not exist for res1. Skipping {res1} {atom1}')  # noqa: E501
+            if True not in two:
+                errmsgs.append(f'{atom2} in PDB does not exist for res2. Skipping {res2} {atom2}')  # noqa: E501
+            
+        except Exception:
+            errmsg = 'Residues and/or atoms are not compatible with PDB file(s)'  # noqa: E501
+        
+        if errmsg is not False:
+            raise SPyCiPDBException(errmsg)
+    
+    errmsgs = [*set(errmsgs)]
+    
+    return format, errmsgs
 
 
-def get_exp_format_pre(fexp):
+def get_exp_format_pre(fexp, fpdb):
     """Get format based on experimental file."""
+    errmsgs = []
     format = {}
     exp = pd.read_csv(fexp)
+    
+    struc = Structure(fpdb)
+    struc.build()
+    struc_residues = struc.data_array[:, col_resSeq]
+    struc_atomnames = struc.data_array[:, col_name]
+    last_residue = int(struc.data_array[:, col_resSeq][-1])
     
     try:
         format['res1'] = exp.res1.values.astype(int).tolist()
@@ -53,13 +102,52 @@ def get_exp_format_pre(fexp):
             )
         raise SPyCiPDBException(errmsg) from err
     
-    return format
+    for i, res1 in enumerate(format['res1']):
+        errmsg = False
+        atom1 = format['atom1'][i]
+        res2 = format['res2'][i]
+        atom2 = format['atom2'][i]
+        
+        if res1 <= 0 or res2 <= 0:
[…]
