+++ src/spycipdb/clis/cli_pre.py
@@ -114,7 +114,7 @@ def main(
     
     log.info(T('reading input paths'))
     pdbs2operate, _istarfile = get_pdb_paths(pdb_files, tmpdir)
-    if len(pdbs2operate) == 0 or pdbs2operate == None:
+    if len(pdbs2operate) == 0 or pdbs2operate is None:
         log.info(
             'No .pdb files were found based on the input. Make sure the '
             'folder/tarball contains .pdb files. Only .tar, .tar.xz, .tar.gz '
@@ -131,7 +131,7 @@ def main(
     execute_pool = pool_function(execute, pdbs2operate, ncores=ncores)
     
     _output = {}
-    _output['format'] = get_exp_format_pre(exp_file)
+    _output['format'], _ = get_exp_format_pre(exp_file, pdbs2operate[0])
     for result in execute_pool:
         _output[result[0].stem] = result[1]
     log.info(S('done'))
