### 2023-01-17 cc567a0
validate resnum for jc
