+++ requirements.yml
@@ -11,6 +11,7 @@ dependencies:
     - pybind11>=2.9.2,<3
     - scipy>=1.8.1,<2
     - matplotlib>=3.5.2,<4
+    - natsort>=8.0.0,<9
     - pip
     - pip:
         - "--editable=git+https://github.com/julie-forman-kay-lab/IDPConformerGenerator.git#egg=idpconfgen"
