### 2023-01-27 99c3aba
fix tox requirements
