+++ requirements.txt
@@ -4,6 +4,6 @@ pybind11>=2.9.2, <3
 pandas>=1.2.4, <2
 scipy>=1.8.1, <2
 matplotlib>=3.5.2, <4
+natsort>=8, <9
 tox>=3.25.0, <4
-natsort>=8.0.0, <9
 git+https://github.com/julie-forman-kay-lab/IDPConformerGenerator
