+++ tox.ini
@@ -36,6 +36,7 @@ deps =
     pytest
     pytest-cov
     hypothesis
+    natsort
 # before running the tests erases any prerecord of coverage
 commands_pre =
     coverage erase
