### 2023-01-27 1ed2622
[SKIP] version bump 0.1.16 -> 0.2.0
[FEATURE] Merge pull request #38 from menoliu/ordered

Output will be ordered as how conformers are on the drive by filename
