### 2023-01-30 #4861 @menoliu
Thank you all for your notes and reviews. I am making improvements to SPyCi-PDB currently as well as one of my projects demand it. I will be modifying the chemical-shift (CS) back-calculator by adding another, more commonly used one, perhaps as the default to hopefully resolve some current/initial issues (I will be more than happy to edit the manuscript if needed later on). Cheers!
