### 2023-03-06 julie-forman-kay-lab/SPyCi-PDB#33 @lohedges
This definitely used to work, but I'm now seeing the following issue on environment creation:

```
Installing pip dependencies: - Ran pip subprocess with arguments:
['/home/lester/.conda/envs/spycipdb/bin/python', '-m', 'pip', 'install', '-U', '-r', '/home/lester/Downloads/JOSS/SPyCi-PDB/condaenv.6x3q1x15.requirements.txt', '--exists-action=b']
Pip subprocess output:
Obtaining idpconfgen from git+https://github.com/julie-forman-kay-lab/IDPConformerGenerator.git#egg=idpconfgen (from -r /home/lester/Downloads/JOSS/SPyCi-PDB/condaenv.6x3q1x15.requirements.txt (line 1))
  Updating ./src/idpconfgen clone
  Installing build dependencies: started
  Installing build dependencies: finished with status 'done'
  Checking if build backend supports build_editable: started
  Checking if build backend supports build_editable: finished with status 'done'
  Getting requirements to build editable: started
  Getting requirements to build editable: finished with status 'error'

Pip subprocess error:
  Running command git fetch -q --tags
  Running command git reset --hard -q eafa437c1f684d819e1bbcbc40313ffea18abdec
  error: subprocess-exited-with-error

  × Getting requirements to build editable did not run successfully.
  │ exit code: 1
  ╰─> [1 lines of output]
      error in idpconfgen setup command: 'python_requires' must be a string containing valid version specifiers; Invalid specifier: '>=3.8.*'
      [end of output]

error: subprocess-exited-with-error

× Getting requirements to build editable did not run successfully.
│ exit code: 1
╰─> See above for output.

failed

CondaEnvException: Pip failed
```

(This occurs both using `mamba` and vanilla `conda`.)

I'll try to debug and let you know if I find a fix.
