### 2023-03-06 julie-forman-kay-lab/SPyCi-PDB#39 @lohedges
[JOSS Review] Warn user if dependencies are missing when environment is updated.
[Re: JOSS Review](https://github.com/openjournals/joss-reviews/issues/4861)

The environment will need to be updated, or recreated, if new dependencies are added. You have recently added `natsort` to the `requirements.txt`, but this is not picked up via `python setup.py develop`, leading to an error when trying to run the code. It would be helpful if the installation (development or otherwise) checked that the requirements were installed before proceeding, or add a note to your documentation to point out that the user should update/recreate the environment first.
