### 2023-03-06 julie-forman-kay-lab/SPyCi-PDB#33 @lohedges
This error occurs if I just try to install `IDPConformerGenerator` directly, i.e.:

```
git clone https://github.com/julie-forman-kay-lab/IDPConformerGenerator
cd IDPConformerGenerator
python setup.py develop
```

Gives:

```
error in idpconfgen setup command: 'python_requires' must be a string containing valid version specifiers; Invalid specifier: '>=3.8.*'
```

Changing to the following works:

```
git diff
diff --git a/setup.py b/setup.py
index 919ffcd..57bff9d 100644
--- a/setup.py
+++ b/setup.py
@@ -83,7 +83,7 @@ setup(
     keywords=[
         # eg: 'keyword1', 'keyword2', 'keyword3',
         ],
-    python_requires='>=3.8.*,<4',
+    python_requires='>=3.8,<4',
     install_requires=[
         ],
     extras_require={
```

Strangely, this file hasn't been updated for four months, so I'm not sure why it's no longer working. However, assuming the change still resolves to the correct version pinning, then I'd suggest updating it. (I'm happy to raise an issue on the other repository if needed.)
