### 2023-03-06 julie-forman-kay-lab/SPyCi-PDB#33 @menoliu
I have updated this in: https://github.com/julie-forman-kay-lab/IDPConformerGenerator/issues/239
