### 2023-03-06 julie-forman-kay-lab/SPyCi-PDB#39 @menoliu
Thanks Lester, that is a good point! I've updated the documentation to reflect that.
