### 2023-03-06 76bf2bd
[SKIP] version bump 0.2.0 -> 0.2.1
Merge pull request #40 from menoliu/warn-upgrade

Warn users when upgrading
