+++ CHANGELOG.rst
@@ -2,6 +2,8 @@
 Changelog
 =========
 
+* Update installation documentation
+
 v0.2.1 (2023-03-06)
 ------------------------------------------------------------
 
