### 2023-03-07 3fc3a34
[SKIP] version bump 0.2.1 -> 0.2.2
Merge pull request #41 from menoliu/update-docs

Update Installation Docs + Temporarily Change IDPConfGen Requirements
