### 2023-03-27 bf88658
update example docs
