+++ CHANGELOG.rst
@@ -2,6 +2,9 @@
 Changelog
 =========
 
+v0.2.3 (2023-03-27)
+------------------------------------------------------------
+
 * Update example usage documentation
 
 v0.2.2 (2023-03-07)
