### 2023-03-28 julie-forman-kay-lab/SPyCi-PDB#43 @JenkeScheen
[JOSS Review] Expand manuscript figure 1 discussion
[Re: JOSS review](https://github.com/openjournals/joss-reviews/issues/4861)

Seeing as the manuscript has minimal results, figure 1 would benefit from an expansion. The author states that panels A and B were generated using `SPyCi-PDB`'s `jc` and `rh` modules (resp.). To showcase `SPyCi-PDB`'s value fully, it is recommended to:
- include results using the remaining modules (or at least the main modules `pre` and `noe`)
- include a short discussion on the interpretation of the figure as well as how it relates to results generated by peers
- [optional] include benchmarking of these outputs compared to peers' outputs (only applicable if results are expected to be different)
