### 2023-04-04 julie-forman-kay-lab/SPyCi-PDB#43 @menoliu
I have generated the figures for NOE and PRE, seeing them now I believe it would be nice to have internal plotting methods in `SPyCi-PDB` as well. I will be creating a PR shortly
