+++ joss/paper.bib
@@ -369,20 +369,6 @@
   journal = {Journal of the American Chemical Society}
 }
 
-@article{Tesei2021,
-  title     = "{DEER-PREdict}: Software for efficient calculation of spin-labeling {EPR} and {NMR} data from conformational ensembles",
-  author    = "Tesei, Giulio and Martins, Jo{\~a}o M and Kunze, Micha B A and Wang, Yong and Crehuet, Ramon and Lindorff-Larsen, Kresten",
-  journal   = "PLoS Comput. Biol.",
-  publisher = "Public Library of Science (PLoS)",
-  volume    =  17,
-  number    =  1,
-  pages     = "e1008551",
-  month     =  jan,
-  year      =  2021,
-  copyright = "http://creativecommons.org/licenses/by/4.0/",
-  language  = "en"
-}
-
 @article{Naudi-Fabra2021,
   title     = "Quantitative description of intrinsically disordered proteins using single-molecule {FRET}, {NMR}, and {SAXS}",
   author    = "Naudi-Fabra, Samuel and Tengo, Maud and Jensen, Malene Ringkj{\o}bing and Blackledge, Martin and Milles, Sigrid",
