### 2023-04-07 4a10260
adds plotting to jc module
