### 2023-04-07 e85034d
adds plotting to rh
