### 2023-04-08 3fc21b9
wrap up and polish plotting
