+++ joss/paper.md
@@ -88,7 +88,9 @@ We plan to integrate alternative methods to calculate experimental datatypes int
 
 Detailed installation/troubleshooting instructions, real-world usage examples, and input/output formats are provided both in the project's documentation hosted on ReadTheDocs (https://spyci-pdb.readthedocs.io/en/stable/) and within the modules through the `--help` argument. Plots of sample outputs from the `jc`, `rh`, `pre`, and `noe` modules using the example structures and data in the repository are shown in \autoref{fig}.
 
-![Plots of distributions of back-calculated experimental data of 100 structures of the unfolded state of the Drk N-terminal SH3 domain (drkN SH3) generated using IDPConformerGenerator [@Teixeira2022]. Panel (A) shows back-calculated <sup>3</sup>J-HNHA couplings in Hz based on the Karplus equation with A, B, and C constants from Lincoff et al. [@Lincoff2020]. Not every residue has a back-calculated J coupling based on the associated experimental file. Panel (B) shows the distribution of back-calculated Rh values in Angstroms using HullRad [@Fleming2018]. Panels (C) and (D) show twenty randomy selected pairs of back-calculated PRE and NOE distances respectively as indicated by the index number corresponding to their experimental counterpart. The ranges of experimental values are represented as grey boxes while back-calculated values are shown as red dots.\label{fig}](fig1.png)
+Comparing the back-calculated PRE and NOE distance values to the experimental observables, the default Euclidean distance interpretation yields some values agreeing with the experimental range (Figure 1C, 1D). With a greater sample size, we would likely capture more back-calculated data agreeing with the experimental ranges. The internal plotting features in the ``noe``, ``pre``, ``rh``, and ``jc`` modules of **SPyCi-PDB** is useful for users to gauge the quality of the initial pool before downstream reweighting. Furthermore, with the integration of different back-calculation methods, these plots will provide the user with a useful comparison between back-calculation philosophies.
+
+![Plots of distributions of back-calculated experimental data of 100 structures of the unfolded state of the Drk N-terminal SH3 domain (drkN SH3) generated using IDPConformerGenerator [@Teixeira2022]. Panel (A) shows back-calculated 3J-HNHA couplings in Hz based on the Karplus equation with A, B, and C constants from Lincoff et al. [@Lincoff2020]. Only residues with experimental data to compare will generate a back-calculated J coupling value. Panel (B) shows the distribution of back-calculated Rh values in Angstroms using HullRad [@Fleming2018]. Panels (C) and (D) show twenty randomly selected pairs of back-calculated PRE and NOE distances, respectively. The ranges of experimental values are represented as grey boxes while back-calculated values for each conformer are shown as red dots.\label{fig}](fig1.png)
 
 # Acknowledgements
 
