### 2023-04-10 1b4d2ef
[SKIP] version bump 0.2.3 -> 0.3.0
[FEATURE] Merge pull request #44 from menoliu/update-paper

Add plotting for `jc`, `rh`, `pre`, and `noe` modules + update manuscript
