+++ CHANGELOG.rst
@@ -2,6 +2,9 @@
 Changelog
 =========
 
+v0.3.0 (2023-04-10)
+------------------------------------------------------------
+
 * Update results in the figure to include PRE and NOE as per #43
 * Add plotting feature to ``rh``, ``jc``, ``pre``, ``noe`` modeules
 
