+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.2.3
+current_version = 0.3.0
 commit = True
 message = 
 	[SKIP] version bump {current_version} -> {new_version}
