### 2023-04-10 #4861 @menoliu
@JenkeScheen Thanks for your review on [#43](https://github.com/julie-forman-kay-lab/SPyCi-PDB/issues/43). I have since updated the manuscript as well as included plotting functions for the modules shown as figures in the paper (`rh`, `jc`, `noe`, `pre`). I have also noticed your checkbox on `Installation instructions:` and have updated the documentation to include a clear list of dependencies as well.
