### 2023-04-12 #4861 @lohedges
Many thanks @menoliu, I can confirm that the fixes work as expected and the installation process is now a lot more streamlined. Longer term, I still think it would be beneficial to make `pip` or `conda` packages for both `SPyCi-PDB` and `IDPConformerGenerator`, as it would make it easier to install your tools into larger environments alongside other packages.

I've left the _references_ section unchecked at present since there is a minor formatting issue. I'll open an issue for that.

@JenkeScheen: Thanks for the comments regarding the additional examples and figures. The updates are really helpful.

Cheers.
