### 2023-04-12 julie-forman-kay-lab/SPyCi-PDB#34 @lohedges
Closing since the error message no longer appears.
