### 2023-04-12 julie-forman-kay-lab/SPyCi-PDB#45 @lohedges
[JOSS Review] Formatting issue in references
[Re: JOSS Review](https://github.com/openjournals/joss-reviews/issues/4861)

A minor syntax error in an italics tag.

![ref_issue](https://user-images.githubusercontent.com/648557/231461805-6a266ee6-3cbd-46da-8dec-e8f7aecca1d1.png)
