### 2023-04-12 julie-forman-kay-lab/SPyCi-PDB#46 @lohedges
[JOSS Review] Explanation of back calculators
[Re: JOSS Review](https://github.com/openjournals/joss-reviews/issues/4861)

Related to [this](https://github.com/julie-forman-kay-lab/SPyCi-PDB/issues/35) issue, it would be nice if the website was updated to provide a little more information regarding the different back calculators that you support. At present there are just acronyms, with little explanation as to why or when you would choose to use one, or what they calculate. (I appreciate that those in the field might know this, but new users might not.) Although brief, this information is provided in the paper so you could simply reproduce it on the website too.

It might be nice to combine the formatting for input/output with the usage section so for each back calculator there is an explanation of what it does, what input is required, and what output it generates. At present you have to jump back-and-forth between sections.
