### 2023-04-12 ba90848
update usage and versioning in paper
