+++ joss/paper.md
@@ -69,7 +69,7 @@ Ultimately, given the complicated and dynamic exchanging nature of IDPs, new bac
 
 As **spycipdb** is written completely in Python, it is compatible with any platform able to execute Python (>=3.8, <4.0). However, certain third-party extensions to perform back-calculations (SAXS and RDC) have only been tested on 64-bit Ubuntu 18.04.X LTS and 20.04.X LTS, as well as WSL 2.0 on 64-bit Windows 11. 
 
-In the production version 0.1.X, four out of eight modules of **SPyCi-PDB**'s back-calculators (`pre`, `noe`, `jc`, `smfret`) use internal mathematical equations and PDB structure processing algorithms from IDPConformerGenerator libraries [@Teixeira2022]. The `pre` (1) and `noe` (2) module calculates scalar distances between pairs of atoms according to the pairs derived from the experimental template. It utilizes an algorithm that matches atom names of each residue with allowance for multiple assignments for `noe`. The `jc` (3) module uses the Karplus curve, a simple cosine function, to back-calculate the desired J-couplings according to residue number as provided by the experimental template file [@Perez2001]. Finally, the `smfret` (4) module takes into consideration residue pairs and a scale factor to adjust for dye size from the experimental setup to back-calculate distances between two alpha-Carbon (CA) atoms [@Lincoff2020]. The aforementioned equations are as follows: 
+In the production version 0.3.X, four out of eight modules of **SPyCi-PDB**'s back-calculators (`pre`, `noe`, `jc`, `smfret`) use internal mathematical equations and PDB structure processing algorithms from IDPConformerGenerator libraries [@Teixeira2022]. The `pre` (1) and `noe` (2) module calculates scalar distances between pairs of atoms according to the pairs derived from the experimental template. It utilizes an algorithm that matches atom names of each residue with allowance for multiple assignments for `noe`. The `jc` (3) module uses the Karplus curve, a simple cosine function, to back-calculate the desired J-couplings according to residue number as provided by the experimental template file [@Perez2001]. Finally, the `smfret` (4) module takes into consideration residue pairs and a scale factor to adjust for dye size from the experimental setup to back-calculate distances between two alpha-Carbon (CA) atoms [@Lincoff2020]. The aforementioned equations are as follows: 
 
 $$
 \begin{aligned}
