+++ docs/usage.rst
@@ -33,6 +33,106 @@ Both will output the help menu.
 .. note::
     All subclients have the :code:`-h` option to show help information.
 
+Formatting for Input and Output Files
+-------------------------------------
+
+Conformers will be required in the PDB format with the ``.pdb`` file extension.
+For tarballs and folders, only the ``.pdb`` files in the folder/tarball will
+be used. Accepted tarballs include ``.tar``, ``.tar.gz``, and ``.tar.xz`` file
+extensions.
+
+Most modules will require a sample experimental results template to base
+the back-calculations off of. Please note that experimental result ``values``
+are not required. For the mathematical equations for the default internal calculators
+(``pre``, ``noe``, ``jc``, ``smfret``) please refer to the paper.
+
+All input files can be in the ``.txt`` file format with comma-delimitation
+for values. Examples can be found in the ``example/drksh3_exp_data/`` folder.
+The required header formatting for the ``.txt`` file for each experimental
+module is highlighted below.
+
+Output files are saved in a standard human-readable ``.JSON`` format.
+In most cases, the first key-value pair gives the format for each of the values
+in subsequent key-value pairs.
+
+Paramagnetic Resonance Entropy (PRE) module
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+| **Input:** ``res1,atom1,res2,atom2``
+| Where res1/atom1 is the residue number and atom name respectively for
+  the first residue and res2/atom2 is the residue number and atom name
+  respectively for the second residue.
+| **Output:** ``'format': {'res1': [], 'atom1': [], 'res2': [], 'atom2': []}``
+| Subsequent keys are the names of the PDB file with a value of: ``[dist_values]``.
+| **About:** The default back-calculator is a simple distance based interpretation of PRE data.
+  Scalar distances between pairs of atoms according to the pairs derived from the experimental template are used.
+
+Nuclear Overhauser Effect (NOE) module
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+| **Input:** ``res1,atom1,atom1_multiple_assignments,res2,atom2,atom2_multiple_assignments``
+| Where res1/atom1 is the residue number and atom name respectively for the first residue and res2/atom2 is 
+  the residue number and atom name respectively for the second residue. 
+  Multiple assignments are either 0 or 1 (no/yes respectively).
+| **Output:** ``'format': {'res1': [], 'atom1': [], 'atom1_multiple_assignments': [], 'res2': [], 'atom2': [], 'atom2_multiple_assignments': []}``
+| Subsequent keys are the names of the PDB file with a value of: ``[dist_values]``.
+| **About:** The default back-calculator is a simple distance based interpretation of NOE data.
+  Scalar distances between pairs of atoms according to the pairs derived from the experimental template are used.
+
+3J-HNHA coupling (JC) module
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+| **Input:** ``resnum``
+| Where ``resnum`` indicates the JC for a specific residue.
+| **Output:** ``'format': [resnum]``
+| Subsequent keys are the names of the PDB file with a value of: ``[jc_values]``.
+| **About:** The default back-calculator uses the Karplus curve, a cosine function, to back-calculate desired
+  J-couplings according to the residue number as provided by the experimental template file.
+
+single-molecule Fluoresence Resonance Energy Transfer (smFRET) module
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+| **Input:** ``res1,res2,scaler``
[…]
