### 2023-04-12 julie-forman-kay-lab/SPyCi-PDB#46 @menoliu
Thanks Lester for getting back to this. That's a great idea I have overhauled the usage section of the documentation so it's a bit more clear
