### 2023-04-12 6e410ea
[SKIP] version bump 0.3.0 -> 0.3.1
Merge pull request #47 from menoliu/update-docs

Add more depth to documentation and fix paper reference artifact
