### 2023-04-12 #4861 @menoliu
Thank you so much @lohedges for the comments. I have addressed them in the latest PR [#47](https://github.com/julie-forman-kay-lab/SPyCi-PDB/pull/47). The documentation is much cleaner now I think.
