### 2023-04-13 b1603b2
[SKIP] version bump 0.3.1 -> 0.3.2
Merge pull request #48 from menoliu/missing-doi

Update Missing DOI and make reference format uniform
