+++ joss/paper.md
@@ -96,4 +96,6 @@ Comparing the back-calculated PRE and NOE distance values to the experimental ob
 
 The motivation behind this project was to create a modular-yet-standalone software package to back-calculate experimental datatypes for conformers generated by the IDPConformerGenerator [@Teixeira2022] platform.
 
+T.H.-G. and J.D.F.-K. acknowledge funding from the National Institute of Health under Grant 2R01GM127627-05. J.D.F.-K. also acknowledges support from the Natural Sciences and Engineering Research Council of Canada (2016-06718) and from the Canada Research Chairs Program.
+
 # References
