+++ CHANGELOG.rst
@@ -2,6 +2,9 @@
 Changelog
 =========
 
+v0.3.3 (2023-04-15)
+------------------------------------------------------------
+
 * Update acknowledgements section of manuscript
 
 v0.3.2 (2023-04-13)
