### 2023-04-26 14445f0
[SKIP] version bump 0.3.3 -> 0.3.4
Merge pull request #50 from menoliu/fix-docs

Fix minor wording in docs
