+++ CHANGELOG.rst
@@ -2,6 +2,9 @@
 Changelog
 =========
 
+v0.3.4 (2023-04-26)
+------------------------------------------------------------
+
 * Minor corrections to documentation
 
 v0.3.3 (2023-04-15)
