### 2023-05-02 julie-forman-kay-lab/SPyCi-PDB#51 @ppxasjsm
Editorial suggestions
Here are some editorial suggestions.
For the statement of need section:
- [x] The sentence "However, interpretation of solution data as a simple calculation from the sum of sampled conformations within IDP/IDR ensembles is fraught with pitfalls" should have a comma after "data" for better clarity.

In the implementation section:
- [x] In the sentence "Thorough testing of each module has been performed...", maybe add a comma after "troubleshooting" for clarity.
- [x] In the sentence "We plan to integrate alternative methods...", maybe add a comma "internally" with "internally, such as" for clarity.
