### 2023-05-02 #4861 @ppxasjsm
@menoliu, I've created a small issue for editorial suggestions of the paper: https://github.com/julie-forman-kay-lab/SPyCi-PDB/issues/51

Please read your publication again and create a new release of the software that reflects the paper version of it. Then deposit this version of the software on Zenodo or Figshare and post the DOI for it here.
