+++ joss/paper.md
@@ -45,7 +45,7 @@ affiliations:
     index: 6
   - name: "Department of Bioengineering, University of California, Berkeley, California 94720-1762, USA"
     index: 7
-date: 30 March 2023
+date: 02 May 2023
 bibliography: paper.bib
 ---
 
@@ -61,7 +61,7 @@ Here we present **SPyCi-PDB**, designed to facilitate and streamline this back-c
 
 # Statement of Need
 
-As new software packages and *in silico* methodologies emerge to better model IDP/IDR structures, back-calculations to multiple experimental datatypes are required to quantitatively assess the conformers generated. However, interpretation of solution data as a simple calculation from the sum of sampled conformations within IDP/IDR ensembles is fraught with pitfalls. For example, commonly used approaches for back-calculating NOE and PRE data for dynamic protein systems treat only the distance and do not incorporate the contribution of dynamics of the vector connecting the interacting points, potentially leading to underestimations of the potential range of distances sampled [@Krzeminski2012; @Brookes2016; @Lincoff2020]. In addition, even for stable systems, back-calculation is not trivial, with even state-of-the-art back-calculators of chemical shifts, such as in UCBShift [@Li2020], leading to errors that can be large relative to the expected deviation of experimental values. Given the rapidly developing nature of different software tools to perform back-calculations, **SPyCi-PDB** should assist by providing a user-friendly, all-in-one package to reduce time and confusion in this back-calculation step as well as open opportunities for future collaborations and integration of new experimental datatypes. Furthermore, **SPyCi-PDB** aims to unify different input and output data formats from different experimental datatypes to increase productivity and accelerate research. As stated in the documentation hosted by ReadTheDocs, input formats are conventional comma-delimited tables (e.g. `.CSV, .TXT`), while the output format is human-readable `.JSON` files that can be easily manipulated using Python or other software based on the user's ultimate needs. **SPyCi-PDB** was also developed to integrate into the IDPConformerGenerator platform [@Teixeira2022].
+As new software packages and *in silico* methodologies emerge to better model IDP/IDR structures, back-calculations to multiple experimental datatypes are required to quantitatively assess the conformers generated. However, interpretation of solution data, as a simple calculation from the sum of sampled conformations within IDP/IDR ensembles is fraught with pitfalls. For example, commonly used approaches for back-calculating NOE and PRE data for dynamic protein systems treat only the distance and do not incorporate the contribution of dynamics of the vector connecting the interacting points, potentially leading to underestimations of the potential range of distances sampled [@Krzeminski2012; @Brookes2016; @Lincoff2020]. In addition, even for stable systems, back-calculation is not trivial, with even state-of-the-art back-calculators of chemical shifts, such as in UCBShift [@Li2020], leading to errors that can be large relative to the expected deviation of experimental values. Given the rapidly developing nature of different software tools to perform back-calculations, **SPyCi-PDB** should assist by providing a user-friendly, all-in-one package to reduce time and confusion in this back-calculation step as well as open opportunities for future collaborations and integration of new experimental datatypes. Furthermore, **SPyCi-PDB** aims to unify different input and output data formats from different experimental datatypes to increase productivity and accelerate research. As stated in the documentation hosted by ReadTheDocs, input formats are conventional comma-delimited tables (e.g. `.CSV, .TXT`), while the output format is human-readable `.JSON` files that can be easily manipulated using Python or other software based on the user's ultimate needs. **SPyCi-PDB** was also developed to integrate into the IDPConformerGenerator platform [@Teixeira2022].
 
[…]
