### 2023-05-02 52f47cb
[SKIP] version bump 0.3.4 -> 0.3.5
Merge pull request #52 from menoliu/update-paper

Update paper with editorial comments
