### 2023-05-10 d537d4f
[SKIP] version bump 0.3.5 -> 0.3.6
Update paper.md
