+++ CHANGELOG.rst
@@ -2,6 +2,9 @@
 Changelog
 =========
 
+v0.3.6 (2023-05-10)
+------------------------------------------------------------
+
 v0.3.5 (2023-05-02)
 ------------------------------------------------------------
 
