+++ CHANGELOG.rst
@@ -2,8 +2,12 @@
 Changelog
 =========
 
+* Update changelog and minor formatting for README
+
 v0.3.6 (2023-05-10)
 ------------------------------------------------------------
+
+* Update paper affiliations
 
 v0.3.5 (2023-05-02)
 ------------------------------------------------------------
