### 2023-05-10 5e97c99
[SKIP] version bump 0.3.6 -> 0.3.7
Merge pull request #53 from menoliu/update_readme

Update readme
