### 2023-05-10 c59080f
[SKIP] version bump 0.3.7 -> 0.3.8
Merge pull request #54 from menoliu/update_paper

Update words in docs for consistency
