+++ CHANGELOG.rst
@@ -2,6 +2,9 @@
 Changelog
 =========
 
+v0.3.8 (2023-05-10)
+------------------------------------------------------------
+
 * Update minor "co-ordinates" to "coordinates" in docs
 
 v0.3.7 (2023-05-10)
