+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.3.7
+current_version = 0.3.8
 commit = True
 message = 
 	[SKIP] version bump {current_version} -> {new_version}
