### 2023-05-10 d34f8f6
[SKIP] version bump 0.3.8 -> 0.3.9
Merge pull request #55 from menoliu/revert-update

Revert versioning and update docs
