+++ CHANGELOG.rst
@@ -2,6 +2,9 @@
 Changelog
 =========
 
+v0.3.9 (2023-05-10)
+------------------------------------------------------------
+
 v0.3.8 (2023-05-10)
 ------------------------------------------------------------
 
