+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.3.8
+current_version = 0.3.9
 commit = True
 message = 
 	[SKIP] version bump {current_version} -> {new_version}
