### 2023-05-10 #4861 @Kevin-Mattheus-Moerman
@menoliu the paper change is fine. On the other changes, should we update the version tag here, e.g. to v0.3.9? And create a new ZENODO archive for it? Or should we proceed with the current ones?
