### 2023-05-10 #4861 @menoliu
@Kevin-Mattheus-Moerman great! I think we should proceed with the current one (v0.3.5) as ReadTheDocs has issues building v0.3.9 (even though I have not changed any code/requirements from v0.3.5 only made edits to `paper.md`).
